Allows to split on keystone door opening, uber state values and fix for ability split and backup saves
Added support to split on keystone being inserted into the door as well as a keystone door opening. Added support for bool, float, int and byte uber states as well comparison support for ==, !=, > and <. Comparison defaults to ==. The uber states are witten as "UberGroupID, UberID, Comparison (optional), Value". I've also fixed abilities splitting when using the in game backup saves menu, this is the only place the Module64 GameAssembly variable is used because I still haven't learned how you get the pointers and all that.